feat: variable channel
This is a rather important feature of jspsych-plus: users should be able
to have an easier way of knowing whether the server / client has made a
change to a variable on their side.

This commit does part of the job: if the server changes an existing
variable, the client would be notified of this change via a web socket
connection.